requirements.txt
Django>=5.0,<5.1
djangorestframework
pyyaml
requests
django-cors-headers
Pillow